feat: add option to restrict search to some directories
Implement the option requested in #24. With `list`, we can now search in
a specific directory and get more accurate results more quickly.

*Note*: unfortunately, this does not seem to work in a performant manner
with `cached_list`, as `locate` does not seem to accept a directory to
search in. The filtering would need to be done in Lua, and that can be
slow.