chore: Remove local tests
They have been broken for a while and the test just start both commands
in telescope. That’s useless without a hight maintenance minimal init
file (nvim just falls back to whatever the user config is). The CI is a
better way to answer the question “Will it work out of the box for a new
user?”. For local tests, it’s best to just manually run the command in a
live nvim instance, as explained in the Readme.