fix: inconsistant return type when looking for a previewer
`find_binary` was returning a string when it did not find a binary in
the table supplied, but a table when it found something. This caused
`table.insert` to error.

`find_binary` now always returns `nil` when it does not find a suitable
executable in the supplied list. This plays quite nicely with the
previews, as when the command supplied is `nil`, we don’t get any
preview. A better behavior might be to use a vim buffer as a previewer,
but that might have performance implications (#42).

Should fix #39 (at least the initial issue)