fix: pass down the opts to the second prompts
Before this patch, the “global” options passed to the Telescope command
were not passed down to the second prompt (the one selecting the files
for instance). This would cause a change of layout set like so:

```
:Telescope repo list layout_strategy=center
```

not to be applied to the second prompt.

This patch fixes that and correctly passes the options down to the
second prompt.

Related to #44 (might fix it)