fix: the second picker couldn't open any file
Due to options being changed for the first picker (in particular, a
custom formatter is set), the second picker was unable to open any file,
it would only point to folders.

This commit fixes this behavior by passing down only the user options to
the second picker, along with the ones specifically needed by it.

Fixes #46