fix: allow cached opts.cwd to be reset if nil
Sometimes, when we cache opts.cwd, it is still nil. While this problem
is investigated [separately][1], it’s worth having this workaround, to
be on the safe side. It’s very disruptive when items can’t be diplayed.

[1]: https://github.com/cljoly/telescope-repo.nvim/pull/70#issuecomment-2014750473

Originally authored by @rish987 in #71. I’ve made cosmetic changes and
added some comments.

Co-authored-by: Rishikesh Vaishnav <rishhvaishnav@gmail.com>